fix: transfer source data to the root folder fix: convert table functional component into class component
