+++ .gitignore
@@ -21,3 +21,8 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# IntellijIdea files
+*.iml
+.idea
+out
