--- pokemon-app/src/components/card-list/card-list.component.jsx
@@ -1,9 +1,0 @@
-import React from "react";
-import {Card} from '../card/card.component';
-import './card-list.styles.css';
-
-export const CardList = (props) => {
-    return <div className='card-list'>{props.pokemons.map(pokemon => (
-        <Card key={pokemon.id} pokemon={pokemon}/>
-    ))}</div>;
-}
