--- pokemon-app/src/components/card-list/card-list.styles.css
@@ -1,27 +1,0 @@
-.card-list{
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-    grid-gap: 2rem;
-
-    width: 85vw;
-
-    margin: 0 auto;
-    padding: 5vh 5vw;
-
-    border: 0.2rem solid #fbd13d;
-    border-radius: 3rem;
-    box-shadow: 0 0 10px rgba(0,0,0,0.5);
-
-    background-color: #d2def1;
-
-    font-family: 'Roboto', sans-serif;
-}
-
-div.card-list h2{
-    font-weight: 300;
-    text-align: center;
-}
-div.card-list h1{
-    font-weight: 500;
-    text-align: center;
-}
