--- pokemon-app/src/components/card/card.styles.css
@@ -1,24 +1,0 @@
-div.card{
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    background-color: rgba(117,124,187,0.3);
-
-    border: 0.2rem solid rgba(117,124,187,0.3);
-    border-radius: 3rem;
-    box-shadow: 0 0 10px rgba(0,0,0,0.5);
-
-    color: #070808;
-    text-shadow: 0.1rem 0.1rem 0.2rem 	#757cbb;
-}
-
-div.card:hover{
-    transform: scale(1.1, 1.1);
-}
-
-div.card img{
-    width: 15vw;
-    height: 25vh;
-}
